refactor: refactor variant core component pattern
