+++ src/ui/components/core/Button/ButtonTypes.ts
@@ -1,9 +1,9 @@
 import { StyleProp, TouchableOpacityProps, ViewStyle } from "react-native"
 import { Text } from "../Text/TextTypes"
-import { variantsKeys } from "./ButtonRoot"
+import { buttonVariantsKeys } from "./ButtonVariants"
 
 export namespace Button {
-	export type VariantsKeys = keyof typeof variantsKeys
+	export type VariantsKeys = keyof typeof buttonVariantsKeys
 	export type Variant = {
 		root: StyleProp<ViewStyle>
 		content: Text.Props
