+++ src/ui/components/core/Text/index.ts
@@ -1,2 +1,3 @@
-export { Text, variantsKeys, getTextVariants } from "./Text"
+export { Text } from "./Text"
 export { type Text as TextTypes } from "./TextTypes"
+export { textVariants, textVariantsKeys } from "./TextVariants"
